Chore: Migrate REST API - getRoomMembers to Typescript (#3899)
