+++ app/lib/rocketchat/rocketchat.js
@@ -309,24 +309,6 @@ const RocketChat = {
 		return sdk.onStreamData(...args);
 	},
 	toggleFavorite,
-	async getRoomMembers({ rid, allUsers, roomType, type, filter, skip = 0, limit = 10 }) {
-		const serverVersion = reduxStore.getState().server.version;
-		if (compareServerVersion(serverVersion, 'greaterThanOrEqualTo', '3.16.0')) {
-			const params = {
-				roomId: rid,
-				offset: skip,
-				count: limit,
-				...(type !== 'all' && { 'status[]': type }),
-				...(filter && { filter })
-			};
-			// RC 3.16.0
-			const result = await sdk.get(`${this.roomTypeToApiType(roomType)}.members`, params);
-			return result?.members;
-		}
-		// RC 0.42.0
-		const result = await this.methodCallWrapper('getUsersOfRoom', rid, allUsers, { skip, limit });
-		return result?.records;
-	},
 	methodCallWrapper(method, ...params) {
 		return sdk.methodCallWrapper(method, ...params);
 	},
